Move `hashtable` and `vector` to `utils/`
